% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{anydate}
\alias{index}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{anytime}{\code{\link[anytime]{anydate}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{zoo}{\code{\link[zoo]{index}}}
}}

