% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_cal}
\alias{tf_cal}
\title{Calculate The TF Factor Value}
\usage{
tf_cal(qt_ptr, name, from_to)
}
\arguments{
\item{qt_ptr}{an R external pointer generated by \code{\link[=tf_quotes_xptr]{tf_quotes_xptr()}}.}

\item{name}{a registered factor name.}

\item{from_to}{a length two \code{Date} vector whose first element must be
smaller than or equal to the first.}
}
\value{
a \code{data.frame} with two columns \emph{DATE} and \emph{VALUE} . The order
of \emph{DATE} column is in ascending order.
}
\description{
Calculate The TF Factor Value
}
