% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_hardware_cores}
\alias{tf_hardware_cores}
\title{The cores on your compture}
\usage{
tf_hardware_cores()
}
\description{
Use \code{std::thread::hardware_concurrency()} to detect the core number on
your compture.
}
