% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_qt_cal}
\alias{tf_qt_cal}
\title{Calculate the TF for single security}
\usage{
tf_qt_cal(qt_ptr, names, from_to)
}
\arguments{
\item{qt_ptr}{An R external pointer generated by \code{\link[=tf_quote_xptr]{tf_quote_xptr()}}.}

\item{names}{A character vector in which all the elements are registered.}

\item{from_to}{A length two \code{Date} vector whose first element must be
smaller than or equal to the first.}
}
\value{
An \code{xts} object in which the columns are the factors' value.
}
\description{
Calculate the TF for single security
}
