% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_qts_cal}
\alias{tf_qts_cal}
\title{Calculate the TF for a security pool}
\usage{
tf_qts_cal(qts_ptr, name, from_to)
}
\arguments{
\item{qts_ptr}{An R external pointer generated by \code{\link[=tf_quotes_xptr]{tf_quotes_xptr()}}.}

\item{name}{a registered factor name}

\item{from_to}{A length two \code{Date} vector whose first element must be
smaller than or equal to the first.}
}
\value{
An \code{xts} object in which the columns are the factor value for
each security in the pool.
}
\description{
Calculate the TF for a security pool
}
