% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_quote_xptr}
\alias{tf_quote_xptr}
\alias{tf_quotes_xptr}
\title{The quote external pointer}
\usage{
tf_quote_xptr(qt_tbl)

tf_quotes_xptr(qt_tbls)
}
\arguments{
\item{qt_tbl}{A \code{data.frame} contains the daily price info
of a security. It must contain a column \code{DATE} and 10 numeric
columns \code{PCLOSE}, \code{OPEN}, \code{HIGH}, \code{LOW}, \code{CLOSE}, \code{VWAP},
\code{VOLUME}, \code{AMOUNT}, \code{BMK_CLOSE}, \code{BMK_OPEN}. Moreover, the
column \code{DATE} must be in ascending order.}
}
\value{
An R external pointer than can hold a \code{Quote} object.
}
\description{
The quote external pointer
}
