% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tf_cal}
\alias{tf_cal}
\title{Calculate the technical factor (TF)}
\usage{
tf_cal(qt_ptr, name, from_to)
}
\arguments{
\item{qt_ptr}{An R external pointer generated by \code{\link[=tf_quote_xptr]{tf_quote_xptr()}}.}

\item{name}{A registered factor name.}

\item{from_to}{A length two \code{Date} vector whose first element must be
smaller than or equal to the first.}
}
\value{
A \code{data.frame} with two columns \code{DATE} and \code{VALUE} . The order
of column \code{DATE} must be in ascending order.
}
\description{
Calculate the technical factor (TF)
}
