% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{tf_quote}
\alias{tf_quote}
\title{The sample quote data used in GCAMCTF}
\source{
The Gildata database
}
\usage{
data(tf_quote)
}
\description{
The data is a realistic China-A security quote of 81 days
in total, from 2018-01-02 ~ 2018-05-07. The column \code{VWAP}
equals to \code{AMOUNT} / \code{VOLUME}. The columns \code{BMK_CLOSE} and
\code{BMK_OPEN} is the close and open price data of the index
\emph{China-A All Shares}.
}
\keyword{data}
